+++ src/utils/validations/date-within-range/date-within-range.js
@@ -1,14 +1,16 @@
 import ValidationsHelper from './../../helpers/validations';
 import DateHelper from './../../helpers/date';
 
-
 // Validates that a date is within a given range
 class DateWithinRangeValidator {
 
   /**
-   * @method constructor
-   * @param {Object} params
-   */
+    * @method constructor
+    * @param {Object} params
+    *   @option [String] customMessage
+    *   @option [Number] limit - number of units
+    *   @option [String] units - unit of time e.g. 'days'/'months'
+    */
   constructor(params = {}) {
     this.customMessage = params.customMessage;
     this.limit = params.limit;
@@ -33,7 +35,7 @@ class DateWithinRangeValidator {
    * @return {String} the error message to display
    */
   message = () => {
-    return ValidationsHelper.validationMessage(this.customMessage, `Date is more than ${this.limit} ${this.units} from today\'s date`);
+    return ValidationsHelper.validationMessage(this.customMessage, 'errors.messages.out_of_range');
   }
 }